refactor: api 폴더를 추가하여 api 관련 로직 분리
