+++ src/hooks/useGetCartItems.ts
@@ -1,26 +1,12 @@
 import { useCallback, useEffect, useState } from 'react';
 import { ProductDTOType } from '../types/product';
+import getCarts from '../api/getCarts';
 
 type cartDataType = {
   id: number;
   quantity: number;
   product: ProductDTOType;
 };
-
-async function getCarts() {
-  const res = await fetch(`${import.meta.env.VITE_API_URL}/cart-items?page=0&size=50`, {
-    headers: {
-      'Content-Type': 'application/json',
-      Authorization: 'Basic cm9zaWVsc2g6cGFzc3dvcmQ',
-    },
-  });
-
-  if (!res.ok) {
-    throw new Error('에러 발생');
-  }
-  const data = await res.json();
-  return data.content;
-}
 
 function useGetCarts() {
   const [carts, setCarts] = useState<cartDataType[] | null>(null);
