+++ src/hooks/useGetProducts.ts
@@ -1,26 +1,6 @@
 import { useEffect, useState } from 'react';
 import { ProductDTOType } from '../types/product';
-import { CATEGORY } from '../constants/products';
-
-type productsOptionType = {
-  category: string;
-  sortKey: string;
-  sortOrder: string;
-};
-
-async function getProducts({ category, sortKey, sortOrder }: productsOptionType) {
-  const res = await fetch(
-    `${import.meta.env.VITE_API_URL}/products?${
-      category !== CATEGORY[0] ? `category=${category}` : ''
-    }&page=0&size=20&sort=${sortKey}%2C${sortOrder}`,
-  );
-
-  if (!res.ok) {
-    throw new Error('에러 발생');
-  }
-  const data = await res.json();
-  return data.content;
-}
+import getProducts from '../api/getProducts';
 
 function useGetProducts({ sort, category }: { sort: string; category: string }) {
   const [products, setProducts] = useState<ProductDTOType[] | null>(null);
